$ yada yada: Initializing the debian directory
Note that yada has some bugs and/or inconvenient "features" in its
changelog generation procedure, so if your project name has a dash,
brace yourself for the warning you see below (Use of uninitialized...)

 $ yada yada
 Creating directory 'debian'
 Use of uninitialized value $1 in substitution iterator at /usr/bin/yada line 4274.
 Creating initial 'debian/changelog'
 Creating example 'debian/packages'

The changelog is a finicky specially formatted file the Debian
toolchain uses for all kinds of things.  Let's fix it.

Signed-off-by: Dan Farina <drfarina@acm.org>