Fix the changelog
We fill in a valid version number, and in the case of ubuntu a valid
distribution name ('unstable' would have been fine on vanilla debian).

A "Valid Version Number" is somewhat involved, depending if you want a
Debian 'native' package or not.  Normally, you don't, so the version
number has to adhere to something like this format:

  x.x.x-n

Such as:

  1.0.0-1

Extreme emphasis on the hyphen, which is the package revision.

This changelog file will have to be maintained by hand or some other
tool; yada does not do it for you (although it does a poor job
generating one to start)

We also change the author of the package to match a local GPG keychain
identity *exactly*.  Generating that identity is beyond the scope of
this document, although Launchpad has reasonably good instructions to
do so:

  https://help.launchpad.net/YourAccount/ImportingYourPGPKey

Signed-off-by: Dan Farina <drfarina@acm.org>