Okay! Time to build.
We're going to use the tool git-buildpackage to assist us in this.  It
is a frontend to dpkg-buildpackage.

git-buildpackage presumes by default there will exist a "commit-ish"
(such a branch) called 'upstream', that represents the version of
software that this debian package is based off of.

The first commit of this project is
f4b43efca98aee4631a055f6bd842d32782c886c, and we'll call that the
"upstream" revision, and make that branch:

  git branch upstream f4b43efca98aee4631a055f6bd842d32782c886c

Once you've done this, you can run "git-buildpackage".  The following
will build, by default, a debian package for your architecture:

 $ git-buildpackage --git-ignore-new

"--git-ignore-new" is just to not stop the tool with an error if there
exist dirty or untracked files.  Liberal application of .gitignore
could let you drop this option, but it does appear its use is somewhat
frequent.

Congratulations! In ../ (above the project directory) you should have
some new files:

  $ cd ../
  $ ls -ld fdr-ppa-demo*
  -rw-r--r-- 1 fdr fdr 5829 2010-12-23 00:09 fdr-ppa-demo_1.0.0-1_amd64.build
  -rw-r--r-- 1 fdr fdr 2001 2010-12-23 00:09 fdr-ppa-demo_1.0.0-1_amd64.changes
  -rw-r--r-- 1 fdr fdr 3104 2010-12-23 00:09 fdr-ppa-demo_1.0.0-1_amd64.deb
  -rw-r--r-- 1 fdr fdr 2818 2010-12-23 00:09 fdr-ppa-demo_1.0.0-1.diff.gz
  -rw-r--r-- 1 fdr fdr 1310 2010-12-23 00:09 fdr-ppa-demo_1.0.0-1.dsc
  -rw-r--r-- 1 fdr fdr 5572 2010-12-22 23:43 fdr-ppa-demo_1.0.0_amd64.build
  -rw-r--r-- 1 fdr fdr 1717 2010-12-22 23:43 fdr-ppa-demo_1.0.0_amd64.changes
  -rw-r--r-- 1 fdr fdr 3100 2010-12-22 23:43 fdr-ppa-demo_1.0.0_amd64.deb
  -rw-r--r-- 1 fdr fdr 1047 2010-12-22 23:43 fdr-ppa-demo_1.0.0.dsc
  -rw-r--r-- 1 fdr fdr  307 2010-12-23 00:09 fdr-ppa-demo_1.0.0.orig.tar.gz
  -rw-r--r-- 1 fdr fdr 2970 2010-12-22 23:43 fdr-ppa-demo_1.0.0.tar.gz

Included in those files is the resultant debian package, compiled and
automatically suffixed with the built architecture.

Signed-off-by: Dan Farina <drfarina@acm.org>